refactor: move session modules and tests to speechlib/tools/ and tests/tools/
- speechlib/batch_process.py        -> speechlib/tools/batch_process.py
- speechlib/extract_unknown_speakers.py -> speechlib/tools/extract_unknown_speakers.py
- tests/test_acceptance_batch_process.py -> tests/tools/
- tests/test_acceptance_extract_unknown_speakers.py -> tests/tools/
- tests/test_speaker_recognition_threshold.py -> tests/tools/
- Fix relative imports: ..core_analysis, ..audio_utils
- Fix patch targets and import paths in tests
- Add speechlib/tools/__init__.py and tests/tools/__init__.py

All 17 tests pass.
