+++ speechlib/tools/process_recordings.py
@@ -13,7 +13,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 sys.path.insert(0, r"c:\workspace\#dev\ClearerVoice-Studio\clearvoice")
 
-from speechlib.batch_process import batch_process
+from speechlib.tools.batch_process import batch_process
 
 VOICES  = Path(r"C:\workspace\#dev\speechlib\transcript_samples\voices")
 UNKNOWN = Path(r"C:\workspace\#dev\speechlib\transcript_samples\voices\_unknown")
